+++ framework/i18n/src/main/java/is/codion/framework/i18n/FrameworkMessages.java
@@ -46,6 +46,7 @@ public final class FrameworkMessages {
 	private static final String UPDATE_TIP = "update_tip";
 	private static final String EDIT = "edit";
 	private static final String EDIT_MNEMONIC = "edit_mnemonic";
+	private static final String EDIT_VALUE_TIP = "edit_value_tip";
 	private static final String EDIT_SELECTED_TIP = "edit_selected_tip";
 	private static final String DELETE = "delete";
 	private static final String DELETE_MNEMONIC = "delete_mnemonic";
@@ -184,6 +185,13 @@ public final class FrameworkMessages {
 	 */
 	public static String editSelectedTip() {
 		return get(EDIT_SELECTED_TIP);
+	}
+
+	/**
+	 * @return edit value tip
+	 */
+	public static String editValueTip() {
+		return get(EDIT_VALUE_TIP);
 	}
 
 	/**